move build earlier
